google::proto::uint64 and quint64 aren't compatible in a 64 bits environnement (unsigned long vs unsigned long long, both 64 bits...).

+++ application/Core/RemoteControlManager/priv/RemoteConnection.cpp
@@ -401,7 +401,7 @@ void RemoteConnection::onNewMessage(Common::MessageHeader::MessageType type, con
          if (currentPassword.isNull() || this->isLocal() || currentPassword == Common::Hash(passMessage.old_password().hash()))
          {
             SETTINGS.set("remote_password", Common::Hash(passMessage.new_password().hash()));
-            SETTINGS.set("salt", passMessage.new_salt());
+            SETTINGS.set("salt", static_cast<quint64>(passMessage.new_salt()));
             SETTINGS.save();
             this->refresh();
          }
